feat: 쿠키 클래스 삭제 후 헤더 값 이용하도록 변경

+++ src/main/java/com/ktb/eatbookappbackend/oauth/jwt/TokenService.java
@@ -32,6 +32,8 @@ public class TokenService {
     }
 
     public void deleteRefreshToken(String refreshToken) {
-        refreshTokenRepository.deleteById(refreshToken);
+        if (refreshToken != null && jwtUtil.validateRefreshToken(refreshToken)) {
+            refreshTokenRepository.deleteById(refreshToken);
+        }
     }
 }
